# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Musician",
    "version": "1.0.0"
  },
  "paths": {
...
    "/create/album": {
      "post": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "message": "álbum criado com sucesso"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "message": "falha ao criar álbum]"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": { "$ref": "#/components/schemas/RequestAlbum" }
            }
          }
        }
      }
    },
    "/create/music/:id": {
      "post": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "message": "música criada com sucesso"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "message": "falha ao criar música"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "ID do álbum a qual a música pertence",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": { "$ref": "#/components/schemas/RequestMusic" }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RequestAlbum": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string",
            "format": "binary"
          },
          "name": {
            "type": "string"
          },
          "authorName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "releaseDate": {
            "type": "string",
            "format": "iso"
          }
        },
        "required": ["music", "image", "name", "authorName", "description", "releaseDate"]
      },
      "RequestMusic": {
        "type": "object",
        "properties": {
          "music": {
            "type": "string",
            "format": "binary"
          },
          "image": {
            "type": "string",
            "format": "binary"
          },
          "name": {
            "type": "string"
          },
          "authorName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "releaseDate": {
            "type": "string",
            "format": "iso"
          }
        },
        "required": ["music", "image", "name", "authorName", "description", "releaseDate"]
      }
    }
  }
}